+++ modules/neuron/include/SpikingNetworks/neuron/IntegrateAndFire.h
@@ -23,7 +23,7 @@ namespace SpikingNetworks::neuron
 		{}
 
 		void integrate(std::vector<SpikingNetworks::core::SpikeEvent> spikes);
-		void fire(float current);
+		void fire(double current);
 
 	protected:
 		double compute(double current);
